attribution: make directive a component
