+++ app/components/attribution.html
@@ -1,5 +1,5 @@
 <i class="fa fa-info-circle"
-  uib-tooltip-template="'html/directives/attribution-tooltip.html'"
+  uib-tooltip-template="$ctrl.templateUrl"
   tooltip-placement="right"
   tooltip-append-to-body="true"
   tooltip-trigger="click"
