+++ app/components/index.ts
@@ -1,5 +1,6 @@
 import about from './about';
 import activity from './activity';
+import attribution from './attribution';
 import authReturn from './auth-return';
 import avatar from './avatar';
 import avatarList from './avatar-list';
@@ -82,6 +83,7 @@ import userProfile from './user-profile';
 export default function(app) {
   about(app);
   activity(app);
+  attribution(app);
   authReturn(app);
   avatar(app);
   avatarList(app);
