+++ package.json
@@ -78,6 +78,7 @@
     "lodash": "^4.3.0",
     "merge-stream": "^1.0.0",
     "mkdirp": "^0.5.1",
+    "ngtemplate-loader": "^1.3.1",
     "nodemon": "^1.8.1",
     "npm-run-all": "^3.0.0",
     "phantomjs-prebuilt": "^2.1.3",
